[refactor] 重構 registerpage & loginpage
